Add build tag nomsgpack (#1852)
* add build tag nomsgpack

* Update copyright

* Update copyright
