+++ .travis.yml
@@ -8,6 +8,9 @@ matrix:
   - go: 1.12.x
     env: GO111MODULE=on
   - go: 1.13.x
+  - go: 1.13.x
+    env: 
+      - TESTTAGS=nomsgpack
   - go: master
 
 git:
